Update 8.0.0.sqlite
This script threw errors when I tried to run it. Most of the fixes are straghtforward, but I am most uncertain of my addition of ajxp_user_prefs to the FROM clause in line 13, as I am not entirely sure what the query is trying to do (this table is empty in my install, so I was able simply to skip this step, but with these changes the query executes without error).